requirements.txt
django==3.2.11
django-environ==0.8.1
djangorestframework==3.12.4
django-cors-headers==3.10.1
django-filter==21.1
django-autoslug==1.9.8
drf-yasg==1.20.0

pillow==9.0.0

argon2-cffi==21.3.0

pytz==2021.3

redis==4.1.0
celery==5.2.3
flower==1.0.0
django-celery-email==3.0.0
django-celery-beat
django-celery-results

djoser==2.1.0
djangorestframework-simplejwt==4.8.0
pyJWT==2.1.0

boto3
django-storages
psycopg2-binary
gunicorn
requests
bs4
type-extensions
lxml
whitenoise
